feat: add shared time utilities
